Rename: ScrollTopBtn.tsx 파일 button 폴더로 이동
